device: allow NM-owned device to be assumed
Team connections should be kept up when NM is stopped so that they can
be assumed upon restart. But now we consider devices created by NM
(those with the @is_nm_owned flag) not capable of assuming connections
and therefore we tear them down and deconfigure when quitting.

Change this and ignore @is_nm_owned when deciding if a device can be
re-assumed.