fixup! dns: use DBus to make dnsmasq nameserver changes
Instead of passing around a GVariantBuilder in NMDnsDnsmasqPrivate,
pass around the final GVariant.

The reason is that the entire argument is constructed at one place,
and later not modified. Passing around a GVariantBuilder makes me
automatically wonder at which places @servers is subsequently modified.
Turns out, nowhere. With the GVariant that is immediately clear because
a variant cannot be modified.