active-connection: add parent active connection tracking
Make it possible to let active connection know about an active
connection it depends on and emit a signal when the parent is active.

XXX: We don't need the parent active connection for the whole lifetime
of an active connection and emit the signal only once. Should we perhaps
replace the setter and signal with a simple callback?