manager: ensure IP config of device is cleared before removal
If the manager removes the device the IP config objects must
be cleared. The reason is that NMPolicy registers to the IP config
changed signal and passes the object on to NMDnsManager.

If the INTERNAL_DEVICE_REMOVED signal is emited with IP configuration
object pending, those objects would be leaked.

This partly redoes the reverted commit f72816bf108cf3fa52ae267e3622a1f19db28b21.

https://bugzilla.gnome.org/show_bug.cgi?id=764483